+++ Controllers/ChambreController.cs
@@ -25,22 +25,23 @@ namespace Hotel.Controllers
         [HttpGet]
         public ActionResult Get()
         {
-            return Ok(Repository.GetAll().Result);
+            //return Ok(Repository.GetAll().Result);
+            return Ok(Repository.GetIncludes(d => d.Categorie));
         }
 
         // GET: api/class/1
-        [HttpGet("{id}")]
+        [HttpGet("{id}", Name = "GetChambre")]
         public ActionResult Get(Guid id)
         {
             if (id.Equals(""))
-                return BadRequest("L'Id reçu en paramètre est invalid");
+                return BadRequest("L'Id reçu en paramètre est invalide");
             else
             {
                 Chambre obj = Repository.GetById(id).Result;
                 if (obj == null)
                     return NotFound("Aucun Objet trouvé avec cet Id");
                 else
-                    return Ok(obj);
+                    return new ObjectResult(obj);
             }
         }
 
@@ -48,31 +49,54 @@ namespace Hotel.Controllers
         [HttpPost]
         public ActionResult Add(Chambre obj)
         {
-            try
+            if (obj == null)
             {
-                Repository.Add(obj);
-            }
-            catch (Exception e)
-            {
-            }
+                return BadRequest();
 
-            return Ok(obj);
+            }
+            Repository.Add(obj);
+
+            return CreatedAtRoute("GetChambre", new { id = obj.Id }, obj);
         }
 
         // PUT: api/class/{obj}
-        [HttpPut]
-        public ActionResult Put(Chambre obj)
+        [HttpPut("{id}")]
+        public ActionResult Put(Guid id, [FromBody] Chambre obj)
         {
+            //Repository.Update(obj);
+            //return Ok(obj);
+
+            if (obj == null || obj.Id != id)
+            {
+                return BadRequest();
+            }
+
+            var temp = Repository.GetById(id).Result;
+            if (temp == null)
+            {
+                return NotFound();
+            }
+
             Repository.Update(obj);
+            //return Ok(Repository.Update(obj).IsCompletedSuccessfully);
             return Ok(obj);
+            //return new NoContentResult();
         }
 
         // DELETE: api/class/{obj}
-        [HttpDelete]
-        public ActionResult Delete(Chambre obj)
+        [HttpDelete("{id}")]
+        public ActionResult Delete(Guid id)
         {
+            //Repository.Delete(obj);
+            //return Ok(obj);
+            var obj = Repository.GetById(id).Result;
+            if (obj == null)
+            {
+                return NotFound();
+            }
+
             Repository.Delete(obj);
-            return Ok(obj);
+            return new NoContentResult();
         }
 
         // GET: api/
