+++ Controllers/ReservationController.cs
@@ -23,22 +23,23 @@ namespace Hotel.Controllers
         [HttpGet]
         public ActionResult Get()
         {
-            return Ok(Repository.GetAll().Result);
+            //return Ok(Repository.GetAll().Result);
+            return Ok(Repository.GetIncludes(d => d.Chambre, e => e.User, f => f.Chambre.Categorie));
         }
 
         // GET: api/class/1
-        [HttpGet("{id}")]
+        [HttpGet("{id}", Name = "GetReservation")]
         public ActionResult Get(Guid id)
         {
             if (id.Equals(""))
-                return BadRequest("L'Id reçu en paramètre est invalid");
+                return BadRequest("L'Id reçu en paramètre est invalide");
             else
             {
                 Reservation obj = Repository.GetById(id).Result;
                 if (obj == null)
                     return NotFound("Aucun Objet trouvé avec cet Id");
                 else
-                    return Ok(obj);
+                    return new ObjectResult(obj);
             }
         }
 
@@ -46,31 +47,57 @@ namespace Hotel.Controllers
         [HttpPost]
         public ActionResult Add(Reservation obj)
         {
-            try
+            if (obj == null)
             {
-                Repository.Add(obj);
+                return BadRequest();
+                
             }
-            catch (Exception e)
-            {
-            }
+            Repository.Add(obj);
 
-            return Ok(obj);
+
+            return CreatedAtRoute("GetReservation", new { id = obj.Id }, obj);
         }
 
         // PUT: api/class/{obj}
-        [HttpPut]
-        public ActionResult Put(Reservation obj)
+        [HttpPut("{id}")]
+        public ActionResult Put(Guid id,[FromBody] Reservation obj)
         {
+
+            //Repository.Update(obj);
+            //return Ok(obj);
+
+            if (obj == null || obj.Id != id)
+            {
+                return BadRequest();
+            }
+
+            var temp = Repository.GetById(id).Result;
+            if (temp == null)
+            {
+                return NotFound();
+            }
             Repository.Update(obj);
+            //return Ok(Repository.Update(obj).IsCompletedSuccessfully);
             return Ok(obj);
+            //return new NoContentResult();
         }
 
         // DELETE: api/class/{obj}
-        [HttpDelete]
-        public ActionResult Delete(Reservation obj)
+        [HttpDelete("{id}")]
+        public ActionResult Delete(Guid id)
         {
+            //Repository.Delete(obj);
+            //return Ok(obj);
+
+
+            var obj = Repository.GetById(id).Result;
+            if (obj == null)
+            {
+                return NotFound();
+            }
+
             Repository.Delete(obj);
-            return Ok(obj);
+            return new NoContentResult();
         }
     }
 }
